+++ src/test/java/web/application/HandlebarsHelperTest.java
@@ -9,7 +9,6 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import web.HandlebarsHelper;
 import web.application.domain.model.User;
 import web.application.dto.UserListResponse;
